rfe449: cast operands to longs so we don't have to worry about the sign bit.
svn path=/autoplot/trunk/QDataSet/; revision=17531

+++ QDataSet/src/org/virbo/dsops/Ops.java
@@ -1855,7 +1855,7 @@ public class Ops {
         return applyBinaryOp(ds1, ds2, new BinaryOp() {
             @Override
             public double op(double d1, double d2) {
-                return (int)d1 & (int)d2;
+                return (long)d1 & (long)d2;
             }
         });
     }
@@ -1875,7 +1875,7 @@ public class Ops {
         return applyBinaryOp(ds1, ds2, new BinaryOp() {
             @Override
             public double op(double d1, double d2) {
-                return (int)d1 | (int)d2;
+                return (long)d1 | (long)d2;
             }
         });
     }
@@ -1897,7 +1897,7 @@ public class Ops {
         return applyBinaryOp(ds1, ds2, new BinaryOp() {
             @Override
             public double op(double d1, double d2) {
-                return (int)d1 ^ (int)d2;
+                return (long)d1 ^ (long)d2;
             }
         });
     }
